chore: reorganise files and update import paths
