+++ packages/backend/src/apps/tiles/actions/find-single-row/index.ts
@@ -2,15 +2,11 @@ import { IRawAction } from '@plumber/types'
 
 import StepError from '@/errors/step'
 import logger from '@/helpers/logger'
-import {
-  getRawRowById,
-  getTableRows,
-  TableRowFilter,
-  TableRowFilterOperator,
-} from '@/models/dynamodb/table-row'
 import Step from '@/models/step'
 import TableCollaborator from '@/models/table-collaborators'
 import TableColumnMetadata from '@/models/table-column-metadata'
+import { getRawRowById, getTableRows } from '@/models/tiles/dynamodb/table-row'
+import { TableRowFilter, TableRowFilterOperator } from '@/models/tiles/types'
 
 import { validateFilters } from '../../common/validate-filters'
 import { FindSingleRowOutput } from '../../types'
