+++ packages/backend/src/graphql/__tests__/mutations/tiles/delete-rows.itest.ts
@@ -2,8 +2,11 @@ import { beforeEach, describe, expect, it } from 'vitest'
 
 import { ForbiddenError } from '@/errors/graphql-errors'
 import deleteRows from '@/graphql/mutations/tiles/delete-rows'
-import { createTableRows, getTableRowCount } from '@/models/dynamodb/table-row'
 import TableMetadata from '@/models/table-metadata'
+import {
+  createTableRows,
+  getTableRowCount,
+} from '@/models/tiles/dynamodb/table-row'
 import User from '@/models/user'
 import Context from '@/types/express/context'
