+++ packages/backend/src/graphql/mutations/tiles/create-rows.ts
@@ -1,6 +1,6 @@
-import { createTableRows } from '@/models/dynamodb/table-row'
 import TableCollaborator from '@/models/table-collaborators'
 import TableMetadata from '@/models/table-metadata'
+import { createTableRows } from '@/models/tiles/dynamodb/table-row'
 
 import type { MutationResolvers } from '../../__generated__/types.generated'
