+++ packages/backend/src/graphql/mutations/tiles/update-row.ts
@@ -1,6 +1,6 @@
-import { updateTableRow } from '@/models/dynamodb/table-row'
 import TableCollaborator from '@/models/table-collaborators'
 import TableMetadata from '@/models/table-metadata'
+import { updateTableRow } from '@/models/tiles/dynamodb/table-row'
 
 import type { MutationResolvers } from '../../__generated__/types.generated'
