+++ packages/backend/src/models/tiles/dynamodb/table-row/index.ts
@@ -1,3 +1,2 @@
 export * from './functions'
 export * from './model'
-export * from './types'
