+++ packages/backend/src/models/tiles/dynamodb/table-row/model.ts
